fix: failing useLoadables test
